- state actions not in list anymore - weighted q-learning finished - grid world example updated

+++ PyPi/approximators/tabular.py
@@ -28,7 +28,7 @@ class Tabular(object):
         assert x.shape[1] == len(self._Q.shape), 'tabular regressor dimension ' \
                                                  'does not fit with input size.'
 
-        for i, idx in enumerate(x.astype(np.int)):
+        for i, idx in enumerate(x):
             self._Q[tuple(idx)] = y[i]
 
     def predict(self, x):
@@ -45,7 +45,7 @@ class Tabular(object):
         assert x.shape[1] == len(self._Q.shape), 'tabular regressor dimension ' \
                                                  'does not fit with input size.'
 
-        return np.array([self._Q[tuple(i)] for i in x.astype(np.int)])
+        return np.array([self._Q[tuple(i)] for i in x])
 
     @property
     def shape(self):
